requirements.txt
Flask
gunicorn
psycopg2-binary
stripe
supabase
python-dotenv